pyproject.toml
[project]
name = "storm"
version = "0.1.0"
description = ""

authors = [
    {name = "Kae-Tempest",email = "katsuhiko.tempest@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "django-ninja",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg[binary] (>=3.2.4,<4.0.0)",
    "django-ninja-jwt (>=5.3.5,<6.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
package-mode = false
